Improve spawn_gtkwave_interactive()
Use '-n +0' option with tail to ensure that it streams the entire VCD dump
file. The previous behavior was racy and dump lines could be lost if tail
was slow to start.

Redirect gtkwave's stdout and stderr to os.devnull instead of closing its
stdout/stderr file descriptors. This seems to improve gtkwave's behavior in
some cases (Python 3.5).

Do not rely on sys.stdin, sys.stdout, and sys.stderr for file descriptors.
When these files have been closed or redirected in the parent process, the
tail, shmidcat, and gtkwave process pipeline needs to be built with the
nominal stdin/stdout/stderr file descriptors.