Exclude event and string from dump_off
These variable types, along with real variables which were already
excluded, cannot take on the 'x' state and thus cannot be represented in a
$dumpoff section.

To accomplish this, VCDWriter now maintains a mapping of Variable instances
to their last value. This replaces the mapping from variable identifiers to
value strings. Having the variable allows it's type to be tested.