Avoid several timestamp duplications
The dumpoff and dumpon paths were a bit sloppy with how/when timemstamps
were dumped. Add new _set_timestamp() and _dump_timestamp() methods along
with _last_dumped_ts state to normalize and generally improve timestamp
handling.

N.B. in the name of performance, _set_timestamp() and _dump_timestamp() are
unrolled in change(). This avoids an approximate 20% performance hit in a
change()-heavy workload (i.e. in the expected workload).