Use black formatting
All code is updated to use black formatting.

https://github.com/psf/black

A pyproject.toml file is added with configuration for the black tool. Note
that skip-string-normalization is enabled; this allows pyvcd code to
continue using single-quoted strings instead of black's default of
double-quoted strings.

The Makefile's 'lint' and 'format' targets are updated to use black.

N.B. there is an issue with installing/running black with pypy3, so instead
of including it in requirements.txt, it is installed conditionally in
.travis.yml.

    https://github.com/psf/black/issues/727
    https://github.com/python/typed_ast/issues/111