Update travis to use bionic
This is a general modernization of .travis.yml.

Use Ubuntu 18.04 "bionic" distribution. Python 2.7, 3.6, 3.7, and 3.8 are
native to travis' bionic dist.

Drop Python 3.4 and 3.5 from the test matrix. These are not in bionic and
pyvcd is dropping support for older Python versions soon.

Use multi-line yaml strings for multi-line script snippets.

Use 'pytest' instead 'py.test'; the former seems to be the canonical name.