Refactor default init value tests
The type-specific default init value selection is moved such that the
default selection and the type test are together in the code. This makes it
much more obvious which default init values apply to which type of
variables.

Some tests are updated to improve coverage of default and non-default init
values.