Repair default string variable value
When calling VCDWriter.register_var() with type='string' and the default
init=None, the string var would be initialized with the value 'x' instead
of '' (the empty string).

This change corrects the default init value for string vars and corrects
the test case that allowed this to pass in the first place.