Compound vector value length correctness
When changing a vector value with a compound size, the length of the value
must match the length of the variable's size tuple. This was not being
enforced and thus it was possible to get unexpected/unwelcome results if a
value of incorrect length was passed to change().

A new test is added to ensure that ValueError is raised for a value/size
length mismatch.

And a test is added to demonstrate that it *is* valid to pass a str of
correct length as a compound value. This is not recommended behavior, but
it was less costly to support this case than to dynamically check and deny
it.