Use base-94 encoded ASCII identifiers
This is an efficient and perfectly hashable encoding for variable
identifier codes. It is based on Tony Bybell's algorithm from
"Implementation of an Efficient Method for Digital Waveform Compression"
which can be found as an appendix in the GTKWave documentation.

The new encoding will make VCD files containing many variables smaller
since each change line will require fewer ascii characters for the
identifier than with the previous base-16 (hexadecimal) encoding.